Add files via upload
public static long fallingPower(int n, int k)
Python has the integer exponentiation operator ** conveniently built in the language, whereas Java unfortunately does not offer that. Exponentiation would be basically useless anyway in a language that uses Yixed size integers that silently hide the overYlows that the integer exponentiation will be producing in spades. (You should also note that in both Python and Java, the caret character ^ de- notes the bitwise exclusive or operation that has nothing to do with integer exponentiation.)
Instead of integer exponentiation, your task is to implement the related operation of falling power that is useful in many combinatorial formulas of discrete calculus. Distinguished syntactically from ordinary powers by underlining the exponent, each term multiplied into the product is always one less than the previous term. For example, the falling power 83 is computed as 8 × 7 × 6 = 336. Simi- larly, the falling power 105 equals 10 × 9 × 8 × 7 × 6 = 30240. Nothing essential changes in this cal- culation when the base is negative; the falling power (–4)5 is computed with the same formula with –4 × –5 × –6 × –7 × –8 = –6720. Similar to normal exponentiation, n0 = 1 for any nonzero integer n.
This method should return the falling power nk where the base n can be any integer, positive or neg- ative or zero, and the exponent k can be any nonnegative integer, so not zero. The JUnit fuzz tests are designed so that your method does not have to worry about potential integer overYlow... provid- ed that you perform your arithmetic calculations with the long kind of 64-bit integers! If you use the bare 32-bit int type anywhere in your method, a silent integer over.low will make yourmethod occasionally mess up its interim values and return incorrect results that fail the JUnit fuzz test. (Yes, even if that method worked when you tried it with your own small values of n and k.)
public static int[] everyOther(int[] arr)
Given an integer array arr, create and return a new array that contains precisely the elements in the even-numbered positions in the array arr. For example, given the array {5, 2, 3, 10, 2}, this method would create and return a new array object {5, 3, 2} that contains the elements in the even-numbered positions 0, 2 and 4 of the original array. Make sure that your method works correctly for arrays of both odd and even lengths, and for empty arrays and singleton arrays with just one element. The length of the result array that you return must also be exactly right so that there are no extra zeros hanging around at the end of that result.
public static int[][] createZigZag(int rows, int cols, int start)
This method creates and returns a new two-dimensional integer array, which in Java is really just a one-dimensional array whose elements are one-dimensional arrays of type int[]. The returned array must have the correct number of rows that each have exactly cols columns. This array must contain the numbers start, start+1, ... , start+(rows*cols-1) in its rows in sorted order, except that the elements in each odd-numbered row must be listed in descending order.
For example, when called with rows=4, cols=5 and start=4, this method should create and re- turn the two-dimensional array whose contents show up as
45678 13 12 11 10 9 14 15 16 17 18 23 22 21 20 19
when displayed in the standard matrix form that is more readable than the more truthful form {{4,5,6,7,8},{13,12,11,10,9},{14,15,16,17,18},{23,22,21,20,19}},in reality a one-dimensional array whose four elements are themselves one-dimensional arrays of integers that serve as rows of this two-dimensional array. In addition to rectangular grids of values, this scheme can represent arbitrary ragged arrays whose rows don’t need to all be of the same length.
public static int countInversions(int[] arr)
Inside an array arr, an inversion is a pair of two positions i<j so that arr[i]>arr[j]. In com- binatorics, the inversion count inside an array is a rough measure of how much “out of order” that array is. If an array is sorted in ascending order, it has zero inversions, whereas an n-element array sorted in reverse order has n(n–1)/2 inversions, the largest number possible. (You will encounter inversions again in this course, if you work through the labs 45 to 47 that deal with permutations and their operations.) This method should count the inversions inside the given array arr, and re- turn that count. As always when writing methods that operate on arrays, make sure that your method works correctly for arrays of any length, including the important special cases of zero and one.